+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `selfupdate` alias for the self update command.
 - PHP 8.2 minimum required.
 - Project dependencies updated.
+- Permissions normalization checks shebang with more patterns.
 
 ## [0.0.4] - 2023-07-03